New Login Page
I made some changes to update the appearances of the login page. These
changes should better integrate the appearances with the rest of the
site. A couple things were left unresolved: the top navigation menu and
an issue I had where the footer on the log in page was hidden - I
didn’t have time to fix that while making updates.